refactor(nav): update navigation to use new SVG logo
